requirements.txt
numpy == 2.4.1
scipy == 1.17.0
matplotlib == 3.10.6
pandas == 2.2.3
seaborn == 0.13.2
tqdm == 4.66.4
torch == 2.5.1
wfdb == 4.2.0
scikit-learn == 1.7.2
vg-beat-detectors == 0.0.13
h5py == 3.13.0
cma == 4.4.1
xgboost == 3.1.3
neurokit2 == 0.2.12
PyYAML == 6.0.1
ipykernel == 6.25.1